+++ gandi/cli/tests/fixtures/api.py
@@ -6,6 +6,8 @@ log = logging.getLogger(__name__)
 
 class Api(object):
 
+    _calls = {}
+
     def request(self, method, apikey, *args, **kwargs):
         log.info('Calling %s%r' % (method, args))
         modname, func = method.split('.', 1)
@@ -13,6 +15,7 @@ class Api(object):
         module = importlib.import_module(modname)
         func = func.replace('.', '_')
         try:
+            self._calls.setdefault(method, []).append(args)
             return getattr(module, func)(*args)
         except Exception as exc:
             log.exception('Unexpected Exception %s while calling %s' % (exc,