+++ CHANGES.rst
@@ -5,6 +5,13 @@ Changelog
 ----
 
 * New command 'gandi domain renew' command to renew a domain.
+* Update 'domain info' command:
+    - add creation, update and expiration date to output
+    - changes nameservers and services output for easier parsing
+* Update 'gandi disk update' command:
+    - add new option '--delete-snapshotprofile' to remove a snapshot
+      profile from disk
+* Fixes #119
 
 0.14
 ----